requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
SQLAlchemy==2.0.35
redis==5.0.8
python-dotenv==1.0.1
aiosqlite==0.19.0
